// frontend/src/client/types.gen.ts
export type AuthorPublic = {
    name: string;
    id: number;
    owner_id?: (string | null);
    books_count?: (number | null);
    created_at?: (string | null);
};
// ...
export type BookPublic = {
    title: string;
    id: number;
    author_id: number;
    owner_id?: (string | null);
    author_name?: (string | null);
    created_at?: (string | null);
};